Group the common checks together
Instead of having the same rules copied for all devices, group the checks that
we run on all of them anyway into a set of checks at the beginning. This way
we can quickly filter out nonmatching devices.

This also removes the ID_INPUT_JOYSTICK tag removal for js devices, the device
should be matched by the preceding rule and thus have the property removed
alrady.

No real functional changes, but we now ignore remove events.